+++ crates/gem_ton/src/signer/transaction/mod.rs
@@ -1,5 +1,7 @@
+mod message;
 mod request;
 mod signing;
+mod wallet;
 #[cfg(test)]
 mod tests;
