+++ crates/gem_ton/testdata/wc_ton_message.json
@@ -1,0 +1,8 @@
+[
+    {
+        "address": "EQAlgB03OjJKdXrlwZiGJD5snSzPKF2VL5bErJn_cqJANGH9",
+        "amount": "241000000",
+        "payload": "{PAYLOAD}",
+        "stateInit": "{PAYLOAD}"
+    }
+]
